+++ src/main/java/net/geforcemods/securitycraft/ClientHandler.java
@@ -41,7 +41,9 @@ import net.geforcemods.securitycraft.screen.IMSScreen;
 import net.geforcemods.securitycraft.screen.InventoryScannerScreen;
 import net.geforcemods.securitycraft.screen.KeyChangerScreen;
 import net.geforcemods.securitycraft.screen.KeycardReaderScreen;
+import net.geforcemods.securitycraft.screen.KeypadBlastFurnaceScreen;
 import net.geforcemods.securitycraft.screen.KeypadFurnaceScreen;
+import net.geforcemods.securitycraft.screen.KeypadSmokerScreen;
 import net.geforcemods.securitycraft.screen.MineRemoteAccessToolScreen;
 import net.geforcemods.securitycraft.screen.ProjectorScreen;
 import net.geforcemods.securitycraft.screen.SCManualScreen;
@@ -123,9 +125,13 @@ public class ClientHandler {
 				"sand",
 				"stone"
 		};
+		ResourceLocation[] furnaces = {
+				new ResourceLocation(SecurityCraft.MODID, "keypad_furnace"),
+				new ResourceLocation(SecurityCraft.MODID, "keypad_smoker"),
+				new ResourceLocation(SecurityCraft.MODID, "keypad_blast_furnace")
+		};
 		//@formatter:on
 		ResourceLocation invScanRL = new ResourceLocation(SecurityCraft.MODID, "inventory_scanner");
-		ResourceLocation ppfRL = new ResourceLocation(SecurityCraft.MODID, "keypad_furnace");
 
 		for (String facing : facings) {
 			for (ResourceLocation facingPoweredBlock : facingPoweredBlocks) {
@@ -137,10 +143,13 @@ public class ClientHandler {
 				registerDisguisedModel(event, facingBlock, "facing=" + facing);
 			}
 
-			registerDisguisedModel(event, ppfRL, "facing=" + facing + ",lit=false,open=false");
-			registerDisguisedModel(event, ppfRL, "facing=" + facing + ",lit=false,open=true");
-			registerDisguisedModel(event, ppfRL, "facing=" + facing + ",lit=true,open=false");
-			registerDisguisedModel(event, ppfRL, "facing=" + facing + ",lit=true,open=true");
+			for (ResourceLocation furnace : furnaces) {
+				registerDisguisedModel(event, furnace, "facing=" + facing + ",lit=false,open=false");
+				registerDisguisedModel(event, furnace, "facing=" + facing + ",lit=false,open=true");
+				registerDisguisedModel(event, furnace, "facing=" + facing + ",lit=true,open=false");
+				registerDisguisedModel(event, furnace, "facing=" + facing + ",lit=true,open=true");
+			}
+
 			registerDisguisedModel(event, invScanRL, "facing=" + facing + ",horizontal=true");
 			registerDisguisedModel(event, invScanRL, "facing=" + facing + ",horizontal=false");
 		}
@@ -211,6 +220,8 @@ public class ClientHandler {
 		RenderTypeLookup.setRenderLayer(SCContent.KEYPAD.get(), cutout);
 		RenderTypeLookup.setRenderLayer(SCContent.KEYPAD_DOOR.get(), cutout);
 		RenderTypeLookup.setRenderLayer(SCContent.KEYPAD_FURNACE.get(), cutout);
+		RenderTypeLookup.setRenderLayer(SCContent.KEYPAD_SMOKER.get(), cutout);
+		RenderTypeLookup.setRenderLayer(SCContent.KEYPAD_BLAST_FURNACE.get(), cutout);
 		RenderTypeLookup.setRenderLayer(SCContent.LASER_BLOCK.get(), cutout);
 		RenderTypeLookup.setRenderLayer(SCContent.LASER_FIELD.get(), translucent);
 		RenderTypeLookup.setRenderLayer(SCContent.REINFORCED_BLACK_STAINED_GLASS.get(), translucent);
@@ -286,6 +297,8 @@ public class ClientHandler {
 		ScreenManager.registerFactory(SCContent.cTypeDisguiseModule, DisguiseModuleScreen::new);
 		ScreenManager.registerFactory(SCContent.cTypeInventoryScanner, InventoryScannerScreen::new);
 		ScreenManager.registerFactory(SCContent.cTypeKeypadFurnace, KeypadFurnaceScreen::new);
+		ScreenManager.registerFactory(SCContent.cTypeKeypadSmoker, KeypadSmokerScreen::new);
+		ScreenManager.registerFactory(SCContent.cTypeKeypadBlastFurnace, KeypadBlastFurnaceScreen::new);
 		ScreenManager.registerFactory(SCContent.cTypeCheckPassword, CheckPasswordScreen::new);
 		ScreenManager.registerFactory(SCContent.cTypeSetPassword, SetPasswordScreen::new);
 		ScreenManager.registerFactory(SCContent.cTypeUsernameLogger, UsernameLoggerScreen::new);
@@ -392,6 +405,8 @@ public class ClientHandler {
 				SCContent.KEYCARD_READER.get(),
 				SCContent.KEYPAD.get(),
 				SCContent.KEYPAD_FURNACE.get(),
+				SCContent.KEYPAD_SMOKER.get(),
+				SCContent.KEYPAD_BLAST_FURNACE.get(),
 				SCContent.LASER_BLOCK.get(),
 				SCContent.PROJECTOR.get(),
 				SCContent.PROTECTO.get(),